V0.1

+++ troc/src/main/java/troc/project/troc/repositories/UserTrocRepository.java
@@ -5,5 +5,5 @@ import troc.project.troc.model.UserTroc;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserTrocRepository extends CrudRepository<UserTroc, Long> {
+public interface UserTrocRepository extends CrudRepository<UserTroc, Integer> {
 }
